+++ crates/rooch-rpc-client/src/lib.rs
@@ -7,7 +7,7 @@ use jsonrpsee::core::client::ClientT;
 use jsonrpsee::http_client::{HttpClient, HttpClientBuilder};
 use moveos_types::{
     function_return_value::FunctionResult, module_binding::MoveFunctionCaller,
-    transaction::FunctionCall, tx_context::TxContext,
+    moveos_std::tx_context::TxContext, transaction::FunctionCall,
 };
 use rooch_client::RoochRpcClient;
 use std::sync::Arc;
