+++ crates/rooch-types/src/framework/address_mapping.rs
@@ -8,9 +8,9 @@ use move_core_types::{account_address::AccountAddress, ident_str, identifier::Id
 use moveos_types::{
     module_binding::{ModuleBinding, MoveFunctionCaller},
     move_option::MoveOption,
+    moveos_std::tx_context::TxContext,
     state::MoveStructState,
     transaction::FunctionCall,
-    tx_context::TxContext,
 };
 
 /// Rust bindings for RoochFramework address_mapping module
