+++ crates/rooch-types/src/framework/auth_validator.rs
@@ -19,9 +19,9 @@ use moveos_types::{
     module_binding::MoveFunctionCaller,
     move_string::MoveAsciiString,
     move_types::FunctionId,
+    moveos_std::tx_context::TxContext,
     state::{MoveStructState, MoveStructType},
     transaction::FunctionCall,
-    tx_context::TxContext,
 };
 use serde::{Deserialize, Serialize};
 use strum_macros::{Display, EnumString};
