+++ crates/rooch-types/src/framework/coin.rs
@@ -7,12 +7,12 @@ use move_core_types::language_storage::StructTag;
 use move_core_types::u256::U256;
 use move_core_types::{account_address::AccountAddress, ident_str, identifier::IdentStr};
 use moveos_types::move_string::MoveString;
-use moveos_types::object::ObjectID;
+use moveos_types::moveos_std::object::ObjectID;
 use moveos_types::state::{MoveStructState, MoveStructType};
 use moveos_types::{
     module_binding::{ModuleBinding, MoveFunctionCaller},
+    moveos_std::tx_context::TxContext,
     transaction::FunctionCall,
-    tx_context::TxContext,
 };
 use serde::{Deserialize, Serialize};
