+++ crates/rooch-types/src/framework/ethereum_validator.rs
@@ -9,9 +9,9 @@ use move_core_types::{
 };
 use moveos_types::{
     module_binding::{ModuleBinding, MoveFunctionCaller},
+    moveos_std::tx_context::TxContext,
     state::MoveStructType,
     transaction::{FunctionCall, MoveAction},
-    tx_context::TxContext,
 };
 
 pub struct EthereumValidator {}
