+++ crates/rooch-types/src/framework/transaction_validator.rs
@@ -12,8 +12,8 @@ use moveos_types::{
     function_return_value::DecodedFunctionResult,
     module_binding::{ModuleBinding, MoveFunctionCaller},
     move_types::FunctionId,
+    moveos_std::tx_context::TxContext,
     transaction::FunctionCall,
-    tx_context::TxContext,
 };
 
 /// Rust bindings for RoochFramework transaction_validator module
