+++ moveos/moveos-store/src/lib.rs
@@ -19,10 +19,10 @@ use crate::state_store::NodeDBStore;
 use crate::transaction_store::{TransactionDBStore, TransactionStore};
 use move_core_types::language_storage::TypeTag;
 use moveos_config::store_config::RocksdbConfig;
-use moveos_types::event::{Event, EventID};
 use moveos_types::event_filter::EventFilter;
 use moveos_types::h256::H256;
-use moveos_types::object::ObjectID;
+use moveos_types::moveos_std::event::{Event, EventID};
+use moveos_types::moveos_std::object::ObjectID;
 use moveos_types::startup_info::StartupInfo;
 use moveos_types::state::State;
 use moveos_types::state_resolver::StateResolver;
