+++ moveos/moveos-store/src/tests/test_state_store.rs
@@ -5,11 +5,11 @@ use crate::MoveOSStore;
 use anyhow::Result;
 use move_core_types::account_address::AccountAddress;
 use move_core_types::effects::{AccountChangeSet, ChangeSet, Op};
-use moveos_types::context;
 use moveos_types::h256::H256;
 use moveos_types::move_string::MoveString;
 use moveos_types::move_types::random_type_tag;
-use moveos_types::object::{NamedTableID, ObjectID};
+use moveos_types::moveos_std::context;
+use moveos_types::moveos_std::object::{NamedTableID, ObjectID};
 use moveos_types::state::{MoveState, State, StateChangeSet, TableChange, TableTypeInfo};
 use rand::{thread_rng, Rng};
 use smt::NodeStore;
