+++ moveos/moveos-store/src/tests/test_store.rs
@@ -10,8 +10,8 @@ use move_core_types::identifier::Identifier;
 use move_core_types::language_storage::{StructTag, TypeTag};
 use move_core_types::vm_status::KeptVMStatus;
 use moveos_config::store_config::RocksdbConfig;
-use moveos_types::event::{Event, EventID};
 use moveos_types::h256::H256;
+use moveos_types::moveos_std::event::{Event, EventID};
 use moveos_types::transaction::TransactionExecutionInfo;
 use raw_store::rocks::{RocksDB, DEFAULT_PREFIX_NAME};
 use raw_store::traits::DBStore;
