+++ moveos/moveos-types/src/lib.rs
@@ -3,8 +3,6 @@
 
 pub mod access_path;
 pub mod addresses;
-pub mod context;
-pub mod event;
 pub mod event_filter;
 pub mod function_return_value;
 pub mod gas_config;
@@ -19,10 +17,8 @@ pub mod move_std;
 pub mod move_string;
 pub mod move_types;
 pub mod moveos_std;
-pub mod object;
 pub mod serde;
 pub mod startup_info;
 pub mod state;
 pub mod state_resolver;
 pub mod transaction;
-pub mod tx_context;
