+++ moveos/moveos-types/src/move_types.rs
@@ -11,7 +11,7 @@ use std::fmt;
 use std::str::FromStr;
 
 use crate::addresses::MOVEOS_STD_ADDRESS;
-use crate::object;
+use crate::moveos_std::object;
 #[cfg(any(test, feature = "fuzzing"))]
 use proptest::prelude::*;
 use rand::prelude::{Distribution, SliceRandom};
