+++ moveos/moveos-types/src/moveos_std/event.rs
@@ -4,8 +4,8 @@
 // Copyright (c) The Diem Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::moveos_std::object::ObjectID;
 use crate::moveos_std::type_info::TypeInfo;
-use crate::object::ObjectID;
 use crate::state::MoveStructType;
 use anyhow::{ensure, Error, Result};
 use move_core_types::account_address::AccountAddress;
