+++ moveos/moveos-types/src/moveos_std/mod.rs
@@ -1,9 +1,13 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod context;
+pub mod event;
 /// This mod contains all the Rust to Move type mapping that are used in MoveosStd
 pub mod module_upgrade_flag;
+pub mod object;
 pub mod object_ref;
+pub mod tx_context;
 pub mod tx_meta;
 pub mod tx_result;
 pub mod type_info;
